Move STI specs to separate file
